analysis/Astrrolok-main/backend/requirements.txt
alembic==1.17.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
arrow==1.4.0
attrs==25.4.0
bcrypt==4.1.3
billiard==4.2.4
black==25.12.0
boto3==1.42.16
botocore==1.42.16
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
faiss-cpu==1.13.2
fastapi==0.110.1
flake8==7.3.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ics==0.7.2
idna==3.11
iniconfig==2.3.0
isort==7.0.0
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jq==1.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==1.2.0
langchain-core==1.2.6
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
librt==0.7.4
lightgbm==4.6.0
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mypy==1.19.1
mypy_extensions==1.1.0
numba==0.63.1
numpy==2.3.5
oauthlib==3.3.1
openai==2.14.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
PyMySQL==1.1.2
pypdf==6.5.0
pyswisseph==2.10.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
scikit-learn==1.8.0
scipy==1.16.3
shap==0.50.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.37.2
TatSu==5.15.1
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.25.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
xxhash==3.6.0
zstandard==0.25.0
